openssh: Use rlimit sandboxing. Using the default seccomp_filter causes sshd to exit with the following error: fatal: privsep_preauth: preauth child terminated by signal 31
This is the second time this issue arose previously it was due to an incompatibility with older Linux kernels. A patch has been merged for that.

This time the cause is unknown, but the workaround that was suggested for the previous issue seems to work now too.

Reference: https://github.com/openssh/openssh-portable/pull/149