asteroid-image-dev: Remove dbg-pkgs dev-pkgs to reduce image size
dbg-pkgs and dev-pkgs add ALL dbg and dev packages to the image. The
result is too large for some watches. And, in particular, adding dev
packages to the rootfs of a smartwatch is of very questionable use,
since this is usually done when development tools such as make, cmake,
gcc etc. are to be run on the target itself. This is unlikely to be
done on a smartwatch.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>