ngfd: Bump SRCREV.
Remove redundant patch file as the g_list_free_full change has been applied upstream now as well.
Use a relative path to to the dbus-gmain.h file as the path to the dbus-gmain isn't actually passed as an include path by autotools.
Reorder the inclusion of autotools and pkgconfig as autotools overrides the build directory which breaks the build as the compiled dbus-gmain library isn't in the expected location (alongside the sources).

Signed-off-by: Darrel Griët <dgriet@gmail.com>