initramfs-scripts-android: Add userdata root to the ext4 search path.
The userdata partition is partially encrypted on newer watches.
In most cases the root of the userdata partition isn't encrypted, but everything in the `media` folder is encrypted.
This adds the root of the userdata partition as a valid location for the ext4, but it means that the ext4 image needs to be pushed externally (i.e. on first boot of the AsteroidOS kernel).

Signed-off-by: Darrel Griët <dgriet@gmail.com>