Update module for mickledore
This updates pulseaduio-modules-nemo to a later version which means it
also moved from autotools to meson.  To allow the project to link
properly, the oe_sysroot option was added to the meson_options.txt file
and a link path changed to use that variable.

The oe_sysroot patch is largely modeled after a similar patch in
meta-openembedded/meta-gnome/recipes-gnome/gnome-control-center by
Andreas Müller.

Signed-off-by: Ed Beroset <beroset@ieee.org>