Add wifi-enabler recipe.
Some Qualcomm based watches need additional steps to enable WiFi.
Ideally this service should be added to the Android init.rc but this does not appear to work since the commands are executed too soon in the boot process.

This provides a systemd service such that WiFi is enabled later in the boot process.

Signed-off-by: Darrel Griët <dgriet@gmail.com>