Remove pkg_postinst_ontarget rule for kernel images.
This should fix the recent device bricks seen. The underlying problem
was identified and described by @FlorentRevest, this is merely an
attempt to translate his observations into a PR.

This means you will not be able to install a new kernel with `opkg`
without also invoking fastboot. OTOH, that never worked reliably.

Some of the removed code was dodgy, and most likely caused
https://github.com/AsteroidOS/meta-asteroid/issues/175

Further code should be removed but this will happen in a separate PR;
this should fix the bricking issue and remove the kernel-update
functionality, creating a consistently empty package for
`linux-$machine-image` is less of a priority.

Fixes https://github.com/AsteroidOS/meta-asteroid/issues/175.