nfcd: Don't autostart service.
NFC seems to cause issues on multiple watches.
On `beluga` it causes for a screen flickering issues on the sides.
On `ray` it causes for random horizontal lines to appear as well as ghost touchscreen presses.

As there's currently little use of NFC disable the service from starting to workaround these issues.

This cause of this issue was first discovered by @argosphil.

Signed-off-by: Darrel Griët <dgriet@gmail.com>