conf: Set OLDEST_KERNEL
This tells Yocto what to pass as --enable-kernel when building glibc.

We need to set this, as the oldest kernel we support is way older
than the 5.15 supported by Yocto.

Setting this option leads to slightly slower code, as it enables
some workarounds which are run even on watches with newer kernels.