dsme: Add patch that fixes (de)serialization and -Werror=format errors
This patch fixes and improves the save/load functions inside DSME that
(de)serialize alarm time. This fixes format specifier related warnings
that otherwise break the build due to -Werror=format-security being set.

This also obviates the need for the workaround in do_compile().

Without this, Yocto Scarthgap builds fail, because the -Werror workaround
in do_compile() breaks compiler calls. The workaround only works if the
-Werror flag is used without any further specificiation (because "-Werror"
is effectively erased cleanly by the sed call); however, if for example
-Werror=format-security is set, the workaround turns this into
"=format-security" (so, not cleanly erasing it), which is not a valid
compiler flag.