Fix geoclue postinstall
This fixes the postinstall step required with geoclue by adding
gsettings to the inherited classes.  That class automatically peforms
the step of compiling the schemas using glib-compile-schemas and also
adds the /usr/share/glib-2.0/schemas to the FILES variable for the
project, which is why it was removed from this file.

The gsettings also has a postinstrm step that recompiles the schemas
after the package is removed.

This fixes https://github.com/AsteroidOS/asteroid/issues/304

Signed-off-by: Ed Beroset <beroset@ieee.org>