libaccounts-glib: Fix and improve recipe
The improvements are:
* Give recipe the proper version number, obviating the need for PV
* Remove unnecessary PR
* Use gobject-introspection bbclass instead of directly depending
  on gobject-introspection recipe
* Replace 0001-meson-Disable-docs-and-tests.patch with backported
  patch that adds "docs" and "test" meson options; that patch is
  also required for the gobject-introspection bbclass
* Use variables (${libdir}, ${datadir}, ${PYTHON_BASEVERSION}) instead
  of hardcoded paths and versions
* Add gettext files to -dev package instead of deleting them (GIR files
  are automatically added to packages by gobject-introspection bbclass)
* Remove unnecessary dependencies (qemu-native) and dependencies that
  are already added by bbclasses (gobject-introspection, vala-native)

Using the gobject-introspection bbclass and the backported patch
allows the recipe to conditionally enable/disable GIR support and
GIR file generation. Previously, GIR files were always generated,
which fails if the OE distro itself has no GIR support enabled.
The bbclass will check for such distro support, and disable GIR
in the recipe if it is not present.

Also, the bbclass sets several important variables that affect
the g-ir-scanner, such as XDG_DATA_DIRS. It also disables the
GIR cache to prevent host contamination (the scanner otherwise
writes cache data to $HOME).

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>