asteroid-machine-config: Rework recipe to fix missing machine.conf
The recipe was most likely causing a race condition where the machine.conf
file was deleted between stages. There is no need to add a special task
though - since the recipe inherits nothing, all that is needed is to
generate the machine.conf directly in do_install ().

Also, cleanup paths to not use hardcoded /etc , and keep the asteroid
config path as a separate variable for clarity.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>